AIX_ML_Models/Drone_Detection_Live_Stream/services/kafka_transmitter/app/requirements.txt
kafka-python
python-socketio==5.7.1
uvicorn[standard]==0.18.3